xe/hw_engine: MMIO base is 32-bit

+++ drivers/gpu/drm/xe/xe_hw_engine.c
@@ -142,7 +142,7 @@ static const struct engine_info engine_infos[] = {
 	},
 };
 
-static uint16_t engine_info_mmio_base(const struct engine_info *info,
+static uint32_t engine_info_mmio_base(const struct engine_info *info,
 				      unsigned int graphics_ver)
 {
 	int i;
